x Buffer_Offset is now int64u for handling repositioning with big atoms
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@704 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze_MinimizeSize.h
@@ -483,7 +483,7 @@ public :
     // Unknown
     //***************************************************************************
 
-    inline void Skip_XX(int64u Bytes, const char*) {Element_Offset+=(size_t)Bytes;}
+    inline void Skip_XX(int64u Bytes, const char*) {Element_Offset+=Bytes;}
 
     //***************************************************************************
     // Flags
